+++ app/controllers/folio/console/api/base_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Folio::Console::Api::BaseController < Folio::ApplicationController
+class Folio::Console::Api::BaseController < Folio::Console::BaseController
   before_action :authenticate_account!
   respond_to :json
 
@@ -9,7 +9,8 @@ class Folio::Console::Api::BaseController < Folio::ApplicationController
   private
 
     def render_error(e)
-      status = Rails.configuration.action_dispatch.rescue_responses[e.class.name] || 500
+      responses = Rails.configuration.action_dispatch.rescue_responses
+      status = responses[e.class.name] || 500
 
       errors = [
         {
@@ -26,12 +27,38 @@ class Folio::Console::Api::BaseController < Folio::ApplicationController
       render json: { data: data }, root: false
     end
 
-    def render_entity(model, entity = nil)
-      if entity.nil?
-        entity = "Entities::Folio::Console::#{model.class.name}".safe_constantize
-        entity ||= "Entities::#{model.class.name}".safe_constantize
-        fail ArgumentError.new('Unknown entity') if entity.nil?
+    def render_record(model, serializer = nil)
+      serializer ||= serializer_for(model)
+
+      if model.valid?
+        render json: serializer.new(model).serializable_hash
+      else
+        render_invalid model
       end
-      render_json entity.represent(model)
     end
+
+    def render_records(models, serializer = nil)
+      serializer ||= serializer_for(models.first)
+      render json: serializer.new(models).serializable_hash
+    end
+
+    def render_invalid(model)
+      errors = model.errors.to_hash.map do |key, msg|
+        title = model.class.human_attribute_name(key)
+
+        { status: 400, title: title, detail: msg }
+      end
+
+      render json: { errors: errors }, status: 400
+    end
+
+    private
+
+      def serializer_for(model)
+        name = model.class.name.gsub('Folio::', '')
+        serializer = "Folio::Console::#{name}Serializer".safe_constantize
+        serializer ||= "#{name}Serializer".safe_constantize
+        fail ArgumentError.new('Unknown serializer') if serializer.nil?
+        serializer
+      end
 end