+++ react/src/ducks/filePlacements.js
@@ -62,7 +62,7 @@ export const filePlacementsSelector = (state) => {
 
   const selected = base.selected.map((filePlacement) => {
     selectedIds.push(filePlacement.id)
-    const file = find(files, { id: filePlacement.file_id })
+    const file = find(files, { id: String(filePlacement.file_id) })
     return {
       ...filePlacement,
       file: file,